requirements.txt
streamlit
langchain
langchain-community
openai
python-dotenv
chromadb
sentence-transformers
pypdf
faiss-cpu
unstructured
tiktoken